Renaming CityRoadsLoader to CityRoads and changing the code according to it.
