+++ routing/CMakeLists.txt
@@ -20,8 +20,8 @@ set(
   checkpoint_predictor.hpp
   checkpoints.cpp
   checkpoints.hpp
-  city_roads_loader.cpp
-  city_roads_loader.hpp
+  city_roads.cpp
+  city_roads.hpp
   city_roads_serialization.hpp
   coding.hpp
   cross_mwm_connector.cpp
